+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Fixed issue with full url of subtitle for HLS data source.
 * Fixed install page from docs.
 * Fixed one of the showcase images
+* [BREAKING_CHANGE] Split controlsHidden into controlsHiddenStart and controlsHiddenEnd.
 
 ## 0.0.77
 * Fixed full screen safe area issue in cupertino controls.